chore(contacts): remove unused PrefillAddAddressFlowRoot
Send already owns the prefilled add-address session, so unmount the
always-on listener roots and drop the unused store/hooks from both apps.
