+++ .changeset/remove-prefill-add-address-root.md
@@ -1,0 +1,7 @@
+---
+"@features/flow-contacts-add-address": patch
+"ledger-live-desktop": patch
+"live-mobile": patch
+---
+
+Remove unused PrefillAddAddressFlowRoot and the shared prefill listener store. Send now owns the prefilled add-address session via startWithPrefilled.
