--- apps/ledger-live-desktop/src/mvvm/features/Contacts/hooks/prefillAddAddress/index.ts
@@ -1,2 +1,0 @@
-export { PrefillAddAddressFlowRoot } from "./PrefillAddAddressFlowRoot";
-export { useOpenPrefillAddAddressFlow } from "@features/flow-contacts-add-address";
