+++ apps/ledger-live-desktop/src/mvvm/features/Contacts/index.ts
@@ -1,3 +1,2 @@
 export { ContactsButton } from "./components/ContactsButton";
 export { default } from "./screens/Contacts";
-export { PrefillAddAddressFlowRoot, useOpenPrefillAddAddressFlow } from "./hooks/prefillAddAddress";
