+++ apps/ledger-live-desktop/src/mvvm/features/GlobalDialogs/index.tsx
@@ -3,7 +3,6 @@ import ModularDialogRoot from "LLD/features/ModularDialog/ModularDialogRoot";
 import SendFlowRoot from "LLD/features/Send/SendFlowRoot";
 import PerpsSignRoot from "LLD/features/Perps/screens/PerpsSign/PerpsSignDialog";
 import ActionConfirmationDialog from "LLD/features/ActionConfirmationDialog";
-import { PrefillAddAddressFlowRoot } from "LLD/features/Contacts";
 
 const ReleaseNotes = lazy(() => import("LLD/features/ReleaseNotes"));
 const BuyDevice = lazy(() => import("LLD/features/BuyDevice"));
@@ -25,7 +24,6 @@ const GlobalDialogs = () => (
     <Suspense fallback={null}>
       <SendFlowRoot />
     </Suspense>
-    <PrefillAddAddressFlowRoot />
     <PerpsSignRoot />
     <ActionConfirmationDialog />
     <Suspense fallback={null}>
