+++ apps/ledger-live-mobile/src/mvvm/features/Contacts/index.ts
@@ -1,4 +1,3 @@
 export { ContactsButton } from "./components/ContactsButton";
 export { ContactsScreen } from "./screens/ContactsPage";
 export type { ContactAddressDetailActionsFlowProps } from "./hooks/useContactAddressDetailActionsAdapter";
-export { PrefillAddAddressFlowRoot, useOpenPrefillAddAddressFlow } from "./hooks/prefillAddAddress";
