+++ features/flow/flow-contacts-add-address/README.md
@@ -8,12 +8,12 @@ Shared Add address flow for Desktop and Mobile.
 ## Scope
 
 - Address validation, selection state, label validation, and flow transitions.
-- Prefilled entry that bypasses currency and address selection and starts on the naming screen.
+- Prefilled entry that bypasses currency and address selection and starts on the naming screen
+  (`startWithPrefilled` on `useAddAddressFlowViewModel`, plus `isPrefillAddAddressFlowOpen`).
 - Dedicated prefilled-address review UI for Web and React Native.
 - Web dialog content and React Native bottom-sheet content.
 - Public ports and helpers used by application-owned currency selection and validation adapters.
-- Shared prefill-session orchestration (`usePrefillAddAddressFlow` and
-  `useOpenPrefillAddAddressFlow`) so Desktop and Mobile only own translations and chrome.
+- Send owns the prefilled add-address session in each app (`useSendPrefillAddAddressFlow`).
 
 ## Structure
 
@@ -24,8 +24,6 @@ Shared Add address flow for Desktop and Mobile.
 - `state/` owns the Add address session, label validation, and currency-selection orchestration.
   Shared Contacts address-entry validation state and input behavior are consumed from
   `@features/platform-contacts`.
-- `usePrefillAddAddressFlow` owns opening, saving, and settling a prefilled session. Apps mount a
-  thin root that renders dialog or drawer chrome from that controller.
 
 ## Public API
