+++ features/flow/flow-contacts-add-address/src/exports.ts
@@ -1,6 +1,4 @@
-export * from "./prefillAddAddress";
-export * from "./useOpenPrefillAddAddressFlow";
-export * from "./usePrefillAddAddressFlow";
+export * from "./isPrefillAddAddressFlowOpen";
 export * from "./state/addressValidation/dependencies";
 export * from "./state/addressValidation/service";
 export * from "./state/ports";
