--- features/flow/flow-contacts-add-address/src/useOpenPrefillAddAddressFlow.ts
@@ -1,8 +1,0 @@
-import type { OpenPrefillAddAddressParams, OpenPrefillAddAddressResult } from "./prefillAddAddress";
-import { requestPrefillAddAddressFlow } from "./prefillAddAddressFlowStore";
-
-export function useOpenPrefillAddAddressFlow(): (
-  params: OpenPrefillAddAddressParams,
-) => Promise<OpenPrefillAddAddressResult> {
-  return requestPrefillAddAddressFlow;
-}
